+++ mapie/estimator/estimator.py
@@ -182,6 +182,7 @@ class EnsembleRegressor(EnsembleEstimator):
         y: ArrayLike,
         train_index: ArrayLike,
         sample_weight: Optional[ArrayLike] = None,
+        **fit_params,
     ) -> RegressorMixin:
         """
         Fit a single out-of-fold model on a given training set.
@@ -204,6 +205,9 @@ class EnsembleRegressor(EnsembleEstimator):
             Sample weights. If None, then samples are equally weighted.
             By default ``None``.
 
+        **fit_params : dict
+            Additional fit parameters.
+
         Returns
         -------
         RegressorMixin
@@ -216,7 +220,11 @@ class EnsembleRegressor(EnsembleEstimator):
             sample_weight = cast(NDArray, sample_weight)
 
         estimator = fit_estimator(
-            estimator, X_train, y_train, sample_weight=sample_weight
+            estimator,
+            X_train,
+            y_train,
+            sample_weight=sample_weight,
+            **fit_params
         )
         return estimator
 
@@ -377,6 +385,7 @@ class EnsembleRegressor(EnsembleEstimator):
         X: ArrayLike,
         y: ArrayLike,
         sample_weight: Optional[ArrayLike] = None,
+        **fit_params,
     ) -> EnsembleRegressor:
         """
         Fit the base estimator under the ``single_estimator_`` attribute.
@@ -396,6 +405,9 @@ class EnsembleRegressor(EnsembleEstimator):
         sample_weight: Optional[ArrayLike] of shape (n_samples,)
             Sample weights. If None, then samples are equally weighted.
             By default ``None``.
+
+        **fit_params : dict
+            Additional fit parameters.
 
         Returns
         -------
@@ -419,7 +431,12 @@ class EnsembleRegressor(EnsembleEstimator):
             )
         else:
             single_estimator_ = self._fit_oof_estimator(
-                clone(estimator), X, y, full_indexes, sample_weight
+                clone(estimator),
+                X,
+                y,
+                full_indexes,
+                sample_weight,
+                **fit_params
             )
             cv = cast(BaseCrossValidator, cv)
             self.k_ = np.full(
@@ -432,7 +449,12 @@ class EnsembleRegressor(EnsembleEstimator):
             else:
                 estimators_ = Parallel(self.n_jobs, verbose=self.verbose)(
                     delayed(self._fit_oof_estimator)(
-                        clone(estimator), X, y, train_index, sample_weight
+                        clone(estimator),
+                        X,
+                        y,
+                        train_index,
+                        sample_weight,
+                        **fit_params
                     )
                     for train_index, _ in cv.split(X)
                 )